requirements.txt
soco~=0.28.1
mfrc522
RPi.GPIO
GitPython
PyYAML~=6.0